Refactored: Bug Fix of singleProductPage
